to continue
the editor I was using kept giving me an error message on my computer and freezing up.  It also kept losing my files and finally I moved to another editor.  That's when I posted to this repo by way of that editor and then git push wouldn't work on the old one.